음성 파일저장/분석 저장 API 작업 완료 - 최선아

+++ src/main/java/com/JobAyong/repository/VoiceRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface VoiceRepository extends JpaRepository<Voice, Long> {
+public interface VoiceRepository extends JpaRepository<Voice, Integer> {
     List<Voice> findByUser(User user);
     List<Voice> findByUserEmail(String email);
 }
